refactor: extract shared evaluation types with re-export shim
Create evaluation.ts with base EvaluationResult interface plus Skill and
Agent extensions. Replace skillEvaluation.ts with a temporary re-export
shim so existing imports keep building during migration.
